Merge initialize_hwloc_topology and cg_initialize_hwloc_topology
Both functions differ only in a call to `read_all_devices` and it was
assumed that if was sufficient to call only one of them. This is not
the case (because `read_all_devices` is required with GPU's and
cgroups).

The difference between both functions in now part of macro's.